+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Support OpenVINO IR (.xml) / ONNX (.onnx) model file for `Explainer` model
 * Enable AISE: Adaptive Input Sampling for Explanation of Black-box Models.
 * Upgrade OpenVINO to 2024.3.0
+* Add saliency map visualization with explanation.plot()
 
 ### What's Changed
 
@@ -17,6 +18,7 @@
 * Try CNN -> ViT assumption for IR insertion by @goodsong81 in https://github.com/openvinotoolkit/openvino_xai/pull/48
 * Enable AISE: Adaptive Input Sampling for Explanation of Black-box Models by @negvet in https://github.com/openvinotoolkit/openvino_xai/pull/49
 * Upgrade OpenVINO to 2024.3.0 by @goodsong81 in https://github.com/openvinotoolkit/openvino_xai/pull/52
+* Add saliency map visualization with explanation.plot() by @GalyaZalesskaya in https://github.com/openvinotoolkit/openvino_xai/pull/53
 
 ### Known Issues